style : style 객체 내부에 선언
